Create speecher1.py
This is my finished program for my capstone project. It is a voice command assistant program called Speecher which runs from the terminal. This program can take a note and save it to a .txt file, perform a Google search, take a screenshot and save it to the downloads folder, and display system information. The program will say the system information as well as display it. I plan on experimenting with this program in the future to see what else I can do.